Implement simple debug API
Note: We du the SPI read/write directly in the communication
setter/getter call here and the setter/getter is blocked as long
as the communication takes.

This makes sense for the debug code, but later on the firmware will
have to do the more high-level logic in the hi3593 tick.